+++ internal/builtins/len.go
@@ -32,13 +32,13 @@ func lenFunction(args []values.Value, ctx values.Ctx) *values.RuntimeResult {
 
 	switch v := value.(type) {
 	case *values.String:
-		return res.Success(values.NewNumber(float64(utf8.RuneCountInString(v.Value))).SetContext(ctx))
+		return res.Success(values.NewNumber(utf8.RuneCountInString(v.Value)).SetContext(ctx))
 	case *values.List:
-		return res.Success(values.NewNumber(float64(len(v.Elements))).SetContext(ctx))
+		return res.Success(values.NewNumber(len(v.Elements)).SetContext(ctx))
 	case *values.Bytes:
-		return res.Success(values.NewNumber(float64(len(v.Data))).SetContext(ctx))
+		return res.Success(values.NewNumber(len(v.Data)).SetContext(ctx))
 	case *values.Map:
-		return res.Success(values.NewNumber(float64(len(v.Keys))).SetContext(ctx))
+		return res.Success(values.NewNumber(len(v.Keys)).SetContext(ctx))
 	default:
 		posStart, posEnd := args[0].GetPos()
