+++ internal/builtins/popen.go
@@ -104,5 +104,5 @@ func popenFunction(args []values.Value, ctx values.Ctx) *values.RuntimeResult {
 	handle := registry.Alloc.Alloc()
 	registry.Processes.Store(handle, &procHandle)
 
-	return res.Success(values.NewNumber(float64(handle)).SetContext(ctx))
+	return res.Success(values.NewNumber(handle).SetContext(ctx))
 }
