+++ internal/builtins/signal.go
@@ -136,7 +136,7 @@ func signalFunction(args []values.Value, ctx values.Ctx) *values.RuntimeResult {
 
 	// Fast path: signal already queued.
 	if sig, ok := tryConsumeSignal(inst); ok {
-		return res.Success(values.NewNumber(float64(signalToInt(sig))).SetContext(ctx))
+		return res.Success(values.NewNumber(signalToInt(sig)).SetContext(ctx))
 	}
 
 	wakeup := make(chan struct{}, 1)
@@ -155,7 +155,7 @@ func signalFunction(args []values.Value, ctx values.Ctx) *values.RuntimeResult {
 	// before signalWaiters saw us, leaving no wakeup queued for us while
 	// signalProbe still reports liveness.
 	if sig, ok := tryConsumeSignal(inst); ok {
-		return res.Success(values.NewNumber(float64(signalToInt(sig))).SetContext(ctx))
+		return res.Success(values.NewNumber(signalToInt(sig)).SetContext(ctx))
 	}
 
 	cancelCh := orch.BeginBlocking(inst, orchestrator.StateBlockedSignal)
@@ -165,13 +165,13 @@ func signalFunction(args []values.Value, ctx values.Ctx) *values.RuntimeResult {
 		select {
 		case <-wakeup:
 			if sig, ok := tryConsumeSignal(inst); ok {
-				return res.Success(values.NewNumber(float64(signalToInt(sig))).SetContext(ctx))
+				return res.Success(values.NewNumber(signalToInt(sig)).SetContext(ctx))
 			}
 
 		case <-cancelCh:
 			// A signal may have landed during the cancel race.
 			if sig, ok := tryConsumeSignal(inst); ok {
-				return res.Success(values.NewNumber(float64(signalToInt(sig))).SetContext(ctx))
+				return res.Success(values.NewNumber(signalToInt(sig)).SetContext(ctx))
 			}
 
 			orch.EndBlocking(inst)
